feat: abstract feelings logic, clean up code
