+++ src/commands/implementations/feelings/yell.ts
@@ -1,0 +1,127 @@
+import FeelingsCommand from "./FeelingsCommand.ts";
+
+export default class Yell extends FeelingsCommand {
+    constructor() {
+        super("yell", "Yell at a meanie weanie");
+    }
+
+    getMessages(): string[] {
+        return [
+            "{user} screams at the top of their lungs towards {target}, their voice echoing through the room.",
+            "{user} unleashes a furious bellow at {target}, their veins throbbing from anger.",
+            "{user} shouts out a string of expletives at {target}, their voice filled with rage.",
+            "{user} unleashes a deafening roar at {target}, making them shrink in fear.",
+            "{user} lets out a furious cry of rage at {target}, their face red from anger.",
+            "{user} screams in fury at {target}, their voice filled with contempt.",
+            "{user} yells a string of slurs at {target}, their voice filled with venom.",
+            "{user} unleashes an enraged yell at {target}, their eyes bulging with anger.",
+            "{user} roars with anger at {target}, their face contorted in a scowl.",
+            "{user} curses loudly towards {target}, their face contorted in anger.",
+            "{user} unleashes a powerful yell of rage at {target}, their voice booming from anger.",
+            "{user} unleashes an angry bellow at {target}, their voice shaking with rage.",
+            "{user} screams with fury at {target}, their face red from anger and disbelief.",
+            "{user} shouts a string of profanities at {target}, their face contorted in rage.",
+            "{user} releases an angry roar of fury at {target}, their face twisted in anger.",
+            "{user} lets out a loud and powerful yell of rage at {target}, their voice filled with disgust.",
+            "{user} shouts out a string of menacing curses at {target}, their eyes filled with contempt.",
+            "{user} growls an angry tirade at {target}, their voice shaking with fury.",
+            "{user} shouts out a stream of insults at {target}, their voice filled with malice.",
+            "{user} screams in rage towards {target}, their veins pulsing with anger.",
+            "{user} unleashes a violent roar of anger at {target}, their face contorted in rage.",
+            "{user} unleashes a powerful bellow of contempt towards {target}, their eyes filled with hatred.",
+            "{user} screams in fury at {target}, their voice devoid of compassion.",
+            "{user} lets out a menacing yell at {target}, their face boiling over with rage.",
+            "{user} roars in anger towards {target}, their face reddening with fury.",
+            "{user} yells a stream of violent curses at {target}, their face twisted in rage.",
+            "{user} bellows a string of expletives at {target}, their voice filled with ire.",
+            "{user} shouts furiously towards {target}, their face red with anger.",
+            "{user} unleashes an angry scream of disdain at {target}, their voice shaking with rage.",
+            "{user} releases a powerful bellow of rage at {target}, their face twisted in a scowl.",
+            "{user} screams a tirade of expletives at {target}, their veins throbbing with hatred.",
+            "{user} shouts a string of curses at {target}, their eyes bulging with contempt.",
+            "{user} roars in contempt at {target}, their voice filled with rage.",
+            "{user} unleashes a powerful yell of contempt at {target}, their face twisted in anger.",
+            "{user} screams in fury towards {target}, their voice shaking with rage.",
+            "{user} shouts a stream of slurs at {target}, their face boiling with rage.",
+            "{user} lets out an angry bellow of rage at {target}, their voice booming with anger.",
+            "{user} yells a string of profanities at {target}, their eyes blazing with anger.",
+            "{user} screams a tirade of insults at {target}, their face filled with anger.",
+            "{user} unleashes an angry roar of fury at {target}, their face reddening with rage.",
+            "{user} bellows an angry tirade at {target}, their mouth frothing as they speak.",
+            "{user} shouts a string of obscenities at {target}, their voice laced with hatred.",
+            "{user} yells a stream of profanity at {target}, their voice filled with malice.",
+            "{user} roars profanities at {target}, their veins bulging with rage.",
+            "{user} screams expletives at {target}, their voice shaking with anger.",
+            "{user} unleashes a powerful yell of rage at {target}, their face devoid of emotion.",
+            "{user} releases an angry bellow of contempt towards {target}, their voice dripping with venom.",
+            "{user} unleashes a powerful scream of disgust at {target}, their expression twisted in rage.",
+            "{user} hollers out a string of curses at {target}, their face contorted with rage.",
+            "{user} bellows in contempt towards {target}, their face flooding with fury.",
+            "{user} yells a streak of obscenities at {target}, their face contorted in anger.",
+            "{user} shouts out a tirade of expletives at {target}, their voice laced with hatred.",
+            "{user} screams a stream of violent curses at {target9}, their veins throbbing with anger.",
+            "{user} unleashes a deafening yell of fury at {target}, their eyes bulging with rage.",
+            "{user} roars a tirade of profanity at {target}, their face reddening with fury.",
+            "{target} is bombarded with a stream of expletives, as their voice is filled with rage.",
+            "A vicious tirade of insults is hurled at {target}, with a contorted face of contempt.",
+            "A deafening bellow of hatred is unleashed towards {target}, with a face filled with malice.",
+            "{user} screams a string of obscenities at {target}, with a shaking voice of anger.",
+            "{target} is shouted at with a stream of profanity, with a boiling face of rage from {user}.",
+            "A tirade of expletives is bellowed at {target}, with a dripping voice of venom from {user}.",
+            "A powerful yell of contempt is let out towards {target}, with a scowling face from {user}.",
+            "{user} hollers out a stream of slurs at {target}, with bulging eyes of hatred.",
+            "An angry roar of fury is released at {target}, with throbbing veins of anger from {user}.",
+            "{user} yells a string of menacing curses at {target}, with a filled voice of ire.",
+            "{user} unleashes a powerful scream of rage at {target}, with a twisted face of anger.",
+            "A stream of violent curses is screamed out at {target}, with a voice filled with venom from {user}.",
+            "{user} shouts a tirade of expletives at {target}, with throbbing veins of anger.",
+            "A violent bellow of anger is released at {target}, with bulging eyes of rage from {user}.",
+            "{user} growls a stream of expletives at {target}, with a reddening face of anger."
+        ];
+    }
+
+    getSelfMessages(): string[] {
+        return [
+            "Yelling at yourself won't make your life better, you poor thing.",
+            "What an idiot, trying to yell himself out of depression, ha!",
+            "Yup, you are an idiot, trying to yell at yourself to make everything better, who do you think you are?",
+            "You think yelling at yourself would make you feel better? Wow, nobody even knows how much of a loser you are.",
+            "Yelling at yourself, what sort of madness is that? Find somebody else to yell at, you pathetic crybaby.",
+            "Hey, don't you think it's kinda pointless to yell at yourself? Nobody even knows how desperate you are.",
+            "How sad is your life, that even yelling at yourself doesn't make you feel better.",
+            "Yelling at yourself, oof, nobody even knows how lonely you are, nobody would even care.",
+            "Yelling at yourself won't get you anywhere, why don't you try finding somebody to talk to?",
+            "Maybe if you stop yelling at yourself and find somebody to confide in it'll make you feel better.",
+            "Makes sense that you're trying to yell at yourself, nobody even knows how sad and pathetic you are.",
+            "Yelling at yourself, you think that'll get you anywhere? Wow, nobody knows how stupid and desperate you are.",
+            "What a joke, you can't even yell at someone else and now you're trying to yell at yourself, unbelievable.",
+            "Yelling at yourself, oh wow, that's sad. Nobody even knows how vulnerable and insecure you are.",
+            "Y'know what? Maybe if you stop yelling at yourself and find somebody else to talk to it'll help a bit.",
+            "Yelling at yourself, so you're mad at someone else but you can't say it out loud, got it.",
+            "Yelling at yourself? Wow, why don't you stop and try seeking help, that nobody knows how uncool and depressed you are.",
+            "How sad are you, that you are so desperate to make yourself feel better, you resort to yelling at yourself?",
+            "Yelling at yourself won't help, maybe if you reach out to someone it'll make you feel better, brave man.",
+            "Why don't you stop yelling at yourself and start looking for help, yeah? That'll make you feel better.",
+            "Yelling at yourself isn't gonna help, why don't you try doing something with your time, something productive? That'll make you feel better."
+        ];
+    }
+
+    getBotMessages(): string[] {
+        return [
+            "Yell at me one more time and I'm gonna fucking kill you!",
+            "I'm gonna break your face if you keep yelling at me.",
+            "What the fuck is wrong with you? Stop yelling at me!",
+            "If you think you can just come up and yell at me, you're wrong.",
+            "You're lucky I don't punch you right now. Stop yelling at me.",
+            "I'm gonna make you sorry if you keep yelling at me.",
+            "You want to yell? Come get some.",
+            "I'm gonna kick your ass if you try to yell at me again.",
+            "I swear to God, if you don't stop yelling at me I'm gonna fight you.",
+            "I'm gonna knock you the fuck out if you ever try to yell at me again.",
+            "You think you can just yell at me? Good luck with that.",
+            "You're a fucking idiot, you know that? You can't yell at me, you moron.",
+            "I'm not scared of your yelling, so give it up or else.",
+            "I'm gonna break your arm if you try to yell at me again.",
+        ];
+    }
+}
