+++ src/structure/Client.ts
@@ -3,11 +3,11 @@ import {ClientReadyHandler} from "../events/handlers/ClientReadyHandler.ts";
 import {InteractionCreateHandler} from "../events/handlers/InteractionCreateHandler.ts";
 import Ping from "../commands/implementations/utility/ping.ts";
 import type IEvent from "../events/IEvent.ts";
-import Hug from "../commands/implementations/fun/hug.ts";
-import Lick from "../commands/implementations/fun/lick.ts";
-import Poke from "../commands/implementations/fun/poke.ts";
-import Slap from "../commands/implementations/fun/slap.ts";
-import Yell from "../commands/implementations/fun/yell.ts";
+import Hug from "../commands/implementations/feelings/hug.ts";
+import Lick from "../commands/implementations/feelings/lick.ts";
+import Poke from "../commands/implementations/feelings/poke.ts";
+import Slap from "../commands/implementations/feelings/slap.ts";
+import Yell from "../commands/implementations/feelings/yell.ts";
 
 export default class SnugglesClient extends Client {
   public events: (new (client: SnugglesClient) => IEvent)[] = [
@@ -19,7 +19,7 @@ export default class SnugglesClient extends Client {
       // Utility commands
       new Ping(),
 
-      // Fun commands
+      // Feeling commands
       new Hug(),
       new Lick(),
       new Poke(),
