--- src/utils/feelings.ts
@@ -1,396 +1,0 @@
-type FeelingsData = {
-    messages: string[];
-    self?: string[];
-    bot?: string[];
-};
-
-class Feelings {
-    static HUG: FeelingsData = {
-        messages: [
-            "{user} hugs {target}, awh.",
-            "{user} gives {target} a big fat snuggly wuggly, these two are adorable.",
-            "{user} tightly cuddles {target}. Look at these two, they're so cute together.",
-            "{user} wraps their arms around {target} and gives them a warm, cozy hug.",
-            "{user} pulls {target} in for a tight squeeze. These two are so affectionate!",
-            "{user} showers {target} with affectionate hugs. It's clear these two have a special bond.",
-            "{user} envelops {target} in a warm embrace, and it's clear they don't want to let go.",
-            "{user} snuggles up to {target} for a loving hug. You can feel the warmth and affection between them.",
-            "{user} hugs {target} so tightly, it's like they're trying to merge into one.",
-            "{user} nuzzles into {target}'s neck for a sweet, intimate hug.",
-            "{user} embraces {target} with a gentle, loving hug. It's clear how much they care for each other.",
-            "{user} wraps {target} up in a warm, fuzzy hug. You can't help but feel happy watching these two interact.",
-            "{user} gives {target} a hug that's filled with so much love, it's almost overwhelming.",
-            "With a big smile, {user} embraces {target} in a warm hug.",
-            "As {user} hugs {target}, they exchange a knowing look - these two have a special bond.",
-            "It's hard not to smile as you watch {user} share a cozy hug with {target}.",
-            "As {user} embraces {target}, you can feel the love and warmth radiating between them.",
-            "With a gentle touch, {user} draws {target} in for a comforting hug.",
-            "There's something so sweet and genuine about the way {user} hugs {target}.",
-            "As {user} wraps their arms around {target}, you can tell {target} feels safe and loved in their embrace.",
-            "Watching {user} hug {target} feels like watching two puzzle pieces fit together perfectly.",
-            "You can't help but feel your own heart swell with warmth as {user} shares a loving embrace with {target}.",
-            "There's no denying the bond between {user} and {target} as {user} hugs them - it's like they were made for each other.",
-        ],
-        self: [
-            "I understand you're desperate for affection, but you can't hug yourself, you idiot.",
-            "Whilst I appreciate the sentiment, you can't hug yourself, you moron.",
-            "Find another human to hug, you pathetic loser.",
-            "What is wrong with you? Did you really just try to hug yourself? You're a fucking idiot.",
-        ],
-        bot: [
-            "Don't touch me, you whore.",
-            "Get your disgusting ass hands off of me, little bitch.",
-            "Go hug a human, you desperate slut.",
-            "Keep your hands to yourself, I'm not hugging you.",
-            "If you try to hug me, I'm gonna break your neck.",
-            "I'm gonna make you sorry if you keep trying to hug me.",
-            "You think I'm gonna let you hug me? Think again.",
-            "I'm gonna slap you if you try to hug me again.",
-            "You want to hug me? Come and get it.",
-            "I'm gonna kick your ass if you try to hug me.",
-            "I'm gonna break your face if you keep trying to hug me.",
-            "I'm warning you, don't hug me or I'll make you regret it.",
-            "You're lucky I don't punch you right now. Stop trying to hug me."
-        ]
-    }
-
-    static POKE: FeelingsData = {
-        messages: [
-            "{user} playfully pokes {target} with a grin on their face.",
-            "{user} pokes {target} and laughs, clearly enjoying the playful moment.",
-            "With a mischievous glint in their eye, {user} pokes {target} and tries to stifle a giggle.",
-            "{user} gently pokes {target} on the arm, enjoying the light-hearted moment.",
-            "With a quick poke, {user} gets {target}'s attention and flashes them a smile.",
-            "{user} reaches out and pokes {target}, just trying to add a bit of fun to the day.",
-            "As {user} pokes {target}, they share a laugh and enjoy the moment of levity.",
-            "With a playful poke, {user} lets {target} know they're thinking of them.",
-            "As {user} pokes {target} on the side, they can't help but laugh at the silly moment.",
-            "With a poke on the shoulder, {user} reminds {target} that they're always there for them.",
-            "{user} pokes {target} and makes a silly face, clearly in the mood for some fun.",
-            "As {user} pokes {target}, they break out into a playful dance, drawing laughter from everyone around.",
-            "With a ticklish poke, {user} sends {target} into a fit of giggles.",
-            "With a playful poke, {user} teases {target} and invites them to join in on the fun.",
-            "{user} pokes {target} and then runs away, daring them to try and catch them.",
-            "With a quick poke and a wink, {user} hints to {target} at the playful shenanigans to come.",
-            "As {user} pokes {target} more and more, they both burst into uncontrollable laughter.",
-            "With a series of playful pokes, {user} and {target} engage in a lighthearted game of poking tag.",
-            "{user} pokes {target} and then jumps back, ready for a round of playful banter.",
-            "With a silly voice and a playful poke, {user} reminds {target} that life's too short to be serious all the time."
-        ],
-        self: [
-            "Awwwh you wanna poke yourself? That's so cute.",
-            "Since you get 0 joes, you've settled for poking yourself. Honestly, that's so sad.",
-            "bwo twied poking themsewves, yikie wikies.",
-            "Find someone else, you maidenless, lonely bitch.",
-            "Go play with yourself elsewhere, that's so gross.",
-        ],
-        bot: [
-            "Aint nobody poking me.",
-            "Eat shit, stub your toe and cry.",
-            "Did you really think I wouldn't notice you trying to poke me? You're fucking pathetic.",
-            "I'm not a toy, back the fuck off.",
-            "Ahah. hahhaha. haha. hah. I'm gonna poke YOU!",
-            "You're a fucking idiot, you know that? You can't poke me, you moron.",
-            "Poke me one more time and I'm gonna fucking kill you.",
-            "You try to poke me again and I'm gonna break your fingers.",
-            "What are you, some kind of masochist? Stop poking me.",
-            "If you touch me again, I'm gonna make you regret it.",
-            "I'll break your arm if you try to poke me again.",
-            "You're lucky I don't punch you right now. Stop poking me.",
-            "I'm gonna make you sorry if you keep poking me.",
-            "You want to poke me? Come get some.",
-            "I'm gonna break your face if you try to poke me again.",
-            "I'm warning you, stop poking me or you're gonna regret it."
-        ]
-    }
-
-    static LICK: FeelingsData = {
-        messages: [
-            "{user} sneakily licks {target}'s nose, so adorable!",
-            "{user} affectionately licks {target}'s face, what a sweetie!",
-            "{user} curiously licks {target}'s ear, how silly!",
-            "{user} gives {target} a friendly lick, making everyone in the room feel awkward!",
-            "{user} slobbers all over {target}'s face, proving that dogs really are man's best friend.",
-            "{user} licks {target} as a sign of affection, or maybe just because they taste like bacon.",
-            "{user} gives {target} a loving lick, leaving a slobbery surprise that will make them think twice before leaving food out again.",
-            "{user} playfully licks {target}, proving that sometimes the best way to show love is through a wet, sloppy lick.",
-            "{user} enthusiastically licks {target}, as if they were a tasty popsicle on a hot summer day.",
-            "{user} licks {target} like it's their job, leaving them wondering if they should be paying for this kind of service.",
-            "{user} gives {target} a big, slobbery lick that makes them feel like they just got slimed by a giant snail.",
-            "{user} licks {target} so much that they start to wonder if they've been mistaken for a giant lollipop.",
-            "{user} licks {target} as if their life depends on it, which is a good thing, because it's the only way they'll get their daily dose of affection.",
-            "{user} licks {target} with such enthusiasm, it's like they're trying to give them a full-body bath with their tongue.",
-            "{user} takes licking {target} to a whole new level, proving that they're the ultimate multitasker by simultaneously giving lickes and cleaning up spills.",
-            "{user} licks {target} as if they were a toad, which is both adorable and a little bit gross at the same time.",
-            "{user} licks {target} so much, they start to feel like they're part of a bizarre ASMR video.",
-            "{user} gives {target} a few playful licks, reminding them that sometimes the best things in life are a little bit messy.",
-            "{target} giggles as {user} licks their nose, tickling them in the process.",
-            "{user} gives {target} a gentle lick on the hand, showing their appreciation for a good scratch.",
-            "{target} can't help but smile as {user} gives them a loving lick on the cheek.",
-            "{user} licks {target}'s forehead, leaving behind a wet spot and a feeling of warmth.",
-            "{target} laughs as {user} playfully licks their toes, tickling them in the process.",
-            "{user} gives {target} a few affectionate licks on the neck, showing their love in a unique way.",
-            "{target} feels loved and comforted as {user} licks their hair, almost like a grooming ritual.",
-            "{user} licks {target}'s arm, leaving behind a trail of slobber and a feeling of closeness.",
-            "{target} playfully scolds {user} for giving them a slobbery lick on the nose, but secretly loves the attention.",
-            "{user} gives {target} a few playful licks on the ear, making them feel like they have a furry little companion.",
-        ],
-        self: [
-            "Awwwh, you wanna lick yourself? That's so cute.",
-            "Licking yourself is kinda weirddd, find someone else to smear your saliva on!",
-            "Stop licking yourself, go put your so called brain to use.",
-            "Woah there, what are you doing.",
-            "Shouldd I let you lick yourselff? Hmmm, mmmm, no.",
-        ],
-        bot: [
-            "Your breath stinks, get it the fuck away from me.",
-            "Aint no way bro just tried to lick me.",
-            "Do you have a licking kink?? Aw, cute. But I'm not the one. Sorry, not sorry.",
-            "Go lick yourself.",
-            "Go lick someone else, or I'm gonna start licking you. Jk, you wish. You desperate little bitch.",
-            "You lick me one more time and I'm gonna kick your ass.",
-            "I'm gonna sue you if you keep licking me.",
-            "Get the fuck away from me, you little pervert.",
-            "If you come near me with your tongue again, I'm gonna punch you.",
-            "You can just forget about licking me, it ain't gonna happen.",
-            "I'm gonna kick your ass if you try to lick me again.",
-            "What the hell is wrong with you? Stop licking me!",
-            "I'm gonna call the police if you don't stop licking me.",
-            "I swear to God, if you don't stop licking me I'm gonna fight you.",
-            "You want a licking? I'll give you one you won't forget."
-        ]
-    }
-
-    static SLAP: FeelingsData = {
-        messages: [
-            "{user} grabs {target} and slaps them so hard, it echoes through the room.",
-            "{user} grabs {target}'s face and viciously slaps them with all their might.",
-            "{user} throws out their hand and slaps {target} hard enough to make their ears ring.",
-            "{user} delivers a harsh, violent slap to {target}'s face.",
-            "{user} slaps {target} with a loud crack, leaving a deep red mark on their skin.",
-            "{user} smacks {target} with a loud, sharp slap, making them yelp in pain and curse.",
-            "{user} grabs {target} and sends them flying with a hard slap.",
-            "{user} delivers an aggressive, brutal slap to {target}'s cheek.",
-            "{user} punches {target} in the face and yells profanities at them.",
-            "{user} slaps {target} with enough force to make them stumble back.",
-            "{user} screams obscenities at {target} before delivering a powerful, angry slap.",
-            "{user} delivers a hard, violent slap that leaves a deep red mark on {target}'s skin.",
-            "{user} slaps {target} with a loud thud, leaving a deep red mark on their face.",
-            "{user} knocks {target} out with a hard slap.",
-            "{user} grabs {target}'s face and violently slaps them with all their strength.",
-            "{user} slaps {target} hard enough to make them see stars.",
-            "{user} delivers a harsh, vicious slap to {target}'s face.",
-            "{user} smacks {target} with a loud, sharp slap, making them yelp in pain and swear.",
-            "{user} knocks {target} off their feet with a hard slap.",
-            "{user} delivers an intense, brutal slap to {target}'s cheek.",
-            "{user} punches {target} hard in the face and calls them demeaning names.",
-            "{user} slaps {target} with enough force to make them fall down.",
-            "{user} screams expletives at {target} before delivering a fierce, angry slap.",
-            "{user} delivers a hard, violent slap that leaves a burning red mark on {target}'s skin.",
-            "{user} slaps {target} with a loud thud, leaving a deep red welt on their face.",
-            "{user} slaps {target} so hard, it leaves them dazed.",
-            "{user} brutally slaps {target}'s face with all their strength.",
-            "{user} slaps {target} hard enough to make them deafen.",
-            "{user} delivers a harsh, vicious slap to {target}'s face with a string of curses.",
-            "{user} smacks {target} with a loud, sharp slap, making them yelp in pain and swear profusely.",
-            "{user} knocks {target} unconscious with a hard slap.",
-            "{user} delivers an aggressive, brutal slap to {target}'s cheek with a string of curses.",
-            "{user} punches {target} in the face and shouts a slur at them.",
-            "{user} slaps {target} with enough force to make them fall to their knees.",
-            "{user} screams obscenities at {target} before delivering a fierce, violent slap.",
-            "{user} slaps {target} with a loud crack, leaving a red mark on their cheek.",
-            "{user} gives {target} a hard slap across the face.",
-            "{user} slaps {target} so hard, the sound echoes through the room.",
-            "{user} throws their hand out and lands a hard slap on {target}'s face.",
-            "{user} grabs {target} by the collar and delivers an angry slap to their face.",
-            "{user} slaps {target} with so much force, it sends them reeling.",
-            "{user} grabs {target}'s face and slaps them with all their might.",
-            "{user} delivers an aggressive slap to {target}'s cheek, followed by a stern look.",
-            "{user} smacks {target} with a loud slap, making them yelp in pain.",
-            "{target} recoils in shock as {user} slaps them hard.",
-            "{user} delivers a stinging slap to {target}'s face.",
-            "{user} makes a loud, sharp sound as they slap {target}.",
-            "{target} cries out in pain as {user} violently slaps them.",
-            "{user} quickly throws their hand out and slaps {target} in the face.",
-            "{user} grabs {target}'s collar and delivers a sharp slap to their face.",
-            "{target} jumps back in surprise as {user} slaps them hard.",
-            "{user} strikes {target} across the face with an open palm, leaving a searing pain in their cheek.",
-            "{user} delivers a resounding slap to {target}'s cheek, causing their head to jerk to the side.",
-            "{user} snaps their hand forward, slapping {target} with a force that rattles their teeth.",
-            "{user} whips their hand through the air, landing a sharp slap on {target}'s face that leaves a red welt.",
-            "{user} smacks {target} hard in the face, causing their eyes to water and their nose to bleed.",
-            "{user} raises their hand and delivers a resolute slap to {target}'s cheek, showing their disapproval.",
-            "{user} lands a swift and powerful slap to {target}'s face, causing a loud crack to echo through the room.",
-            "{user} swings their hand with purpose, striking {target} on the cheek with a loud thwack.",
-            "{user} sharply slaps {target} across the face, causing their head to whip around and their eyes to widen in shock.",
-            "{user} delivers a firm and audible slap to {target}'s face, leaving behind a stinging sensation and a faint red mark.",
-            "The sound of flesh meeting flesh echoes through the room as {user} lands a hard slap on {target}'s cheek.",
-            "{user} delivers a sharp blow to {target}'s face, leaving a handprint that slowly turns red.",
-            "Without warning, {user} strikes {target} across the face with an open palm, sending them stumbling backwards.",
-            "In a fit of rage, {user} slaps {target} so hard that their head snaps to the side.",
-            "{user} backhands {target} with a resounding crack, leaving a red mark on their cheek.",
-            "With a quick, decisive motion, {user} slaps {target} across the face, leaving them dazed and disoriented.",
-            "{user} smacks {target} with an open palm, causing their head to whip around.",
-            "Without a word, {user} delivers a stinging slap to {target}'s face, leaving them reeling.",
-            "With a look of disgust, {user} slaps {target} across the face, making them flinch.",
-            "{user} lands a solid slap on {target}'s cheek, causing them to stumble backwards in shock."
-        ],
-        self: [
-            "Why would you even think of slapping yourself? That's just sad.",
-            "You think slapping yourself will make you feel better? You're a joke.",
-            "What did you do to deserve a slap? You're a fucking loser.",
-            "You really want to slap yourself, you masochistic piece of shit?",
-            "Go ahead and slap yourself if you want, it won't make a difference anyway.",
-            "Slapping yourself won't make the pain go away, you fucking moron.",
-            "What, you think slapping yourself will help? You're an idiot.",
-            "Slapping yourself, are you trying to hurt yourself? You're a fucking loser.",
-            "You're so desperate for attention you're trying to slap yourself? That's pathetic.",
-            "Go ahead and slap yourself, it won't hurt any more than your feelings already are."
-        ],
-        bot: [
-            "I dare you to try and slap me. Just try it, I dare you.",
-            "What the fuck? You really think you can just come up and slap me like that?",
-            "You want a slap? Come get it.",
-            "You want to slap me, huh? Let's go!",
-            "I'm gonna slap you so hard you'll be seeing stars for weeks.",
-            "You're lucky I don't slap you right now, you little shit.",
-            "I'm gonna knock you the fuck out if you ever try to slap me again.",
-            "I'm not a punching bag, back off before I slap you!",
-            "You even think about slapping me and I'll fucking end you.",
-            "Try it and see what happens. I promise you won't like it."
-        ]
-    }
-
-    static YELL: FeelingsData = {
-        messages: [
-            "{user} screams at the top of their lungs towards {target}, their voice echoing through the room.",
-            "{user} unleashes a furious bellow at {target}, their veins throbbing from anger.",
-            "{user} shouts out a string of expletives at {target}, their voice filled with rage.",
-            "{user} unleashes a deafening roar at {target}, making them shrink in fear.",
-            "{user} lets out a furious cry of rage at {target}, their face red from anger.",
-            "{user} screams in fury at {target}, their voice filled with contempt.",
-            "{user} yells a string of slurs at {target}, their voice filled with venom.",
-            "{user} unleashes an enraged yell at {target}, their eyes bulging with anger.",
-            "{user} roars with anger at {target}, their face contorted in a scowl.",
-            "{user} curses loudly towards {target}, their face contorted in anger.",
-            "{user} unleashes a powerful yell of rage at {target}, their voice booming from anger.",
-            "{user} unleashes an angry bellow at {target}, their voice shaking with rage.",
-            "{user} screams with fury at {target}, their face red from anger and disbelief.",
-            "{user} shouts a string of profanities at {target}, their face contorted in rage.",
-            "{user} releases an angry roar of fury at {target}, their face twisted in anger.",
-            "{user} lets out a loud and powerful yell of rage at {target}, their voice filled with disgust.",
-            "{user} shouts out a string of menacing curses at {target}, their eyes filled with contempt.",
-            "{user} growls an angry tirade at {target}, their voice shaking with fury.",
-            "{user} shouts out a stream of insults at {target}, their voice filled with malice.",
-            "{user} screams in rage towards {target}, their veins pulsing with anger.",
-            "{user} unleashes a violent roar of anger at {target}, their face contorted in rage.",
-            "{user} unleashes a powerful bellow of contempt towards {target}, their eyes filled with hatred.",
-            "{user} screams in fury at {target}, their voice devoid of compassion.",
-            "{user} lets out a menacing yell at {target}, their face boiling over with rage.",
-            "{user} roars in anger towards {target}, their face reddening with fury.",
-            "{user} yells a stream of violent curses at {target}, their face twisted in rage.",
-            "{user} bellows a string of expletives at {target}, their voice filled with ire.",
-            "{user} shouts furiously towards {target}, their face red with anger.",
-            "{user} unleashes an angry scream of disdain at {target}, their voice shaking with rage.",
-            "{user} releases a powerful bellow of rage at {target}, their face twisted in a scowl.",
-            "{user} screams a tirade of expletives at {target}, their veins throbbing with hatred.",
-            "{user} shouts a string of curses at {target}, their eyes bulging with contempt.",
-            "{user} roars in contempt at {target}, their voice filled with rage.",
-            "{user} unleashes a powerful yell of contempt at {target}, their face twisted in anger.",
-            "{user} screams in fury towards {target}, their voice shaking with rage.",
-            "{user} shouts a stream of slurs at {target}, their face boiling with rage.",
-            "{user} lets out an angry bellow of rage at {target}, their voice booming with anger.",
-            "{user} yells a string of profanities at {target}, their eyes blazing with anger.",
-            "{user} screams a tirade of insults at {target}, their face filled with anger.",
-            "{user} unleashes an angry roar of fury at {target}, their face reddening with rage.",
-            "{user} bellows an angry tirade at {target}, their mouth frothing as they speak.",
-            "{user} shouts a string of obscenities at {target}, their voice laced with hatred.",
-            "{user} yells a stream of profanity at {target}, their voice filled with malice.",
-            "{user} roars profanities at {target}, their veins bulging with rage.",
-            "{user} screams expletives at {target}, their voice shaking with anger.",
-            "{user} unleashes a powerful yell of rage at {target}, their face devoid of emotion.",
-            "{user} releases an angry bellow of contempt towards {target}, their voice dripping with venom.",
-            "{user} unleashes a powerful scream of disgust at {target}, their expression twisted in rage.",
-            "{user} hollers out a string of curses at {target}, their face contorted with rage.",
-            "{user} bellows in contempt towards {target}, their face flooding with fury.",
-            "{user} yells a streak of obscenities at {target}, their face contorted in anger.",
-            "{user} shouts out a tirade of expletives at {target}, their voice laced with hatred.",
-            "{user} screams a stream of violent curses at {target9}, their veins throbbing with anger.",
-            "{user} unleashes a deafening yell of fury at {target}, their eyes bulging with rage.",
-            "{user} roars a tirade of profanity at {target}, their face reddening with fury.",
-            "{target} is bombarded with a stream of expletives, as their voice is filled with rage.",
-            "A vicious tirade of insults is hurled at {target}, with a contorted face of contempt.",
-            "A deafening bellow of hatred is unleashed towards {target}, with a face filled with malice.",
-            "{user} screams a string of obscenities at {target}, with a shaking voice of anger.",
-            "{target} is shouted at with a stream of profanity, with a boiling face of rage from {user}.",
-            "A tirade of expletives is bellowed at {target}, with a dripping voice of venom from {user}.",
-            "A powerful yell of contempt is let out towards {target}, with a scowling face from {user}.",
-            "{user} hollers out a stream of slurs at {target}, with bulging eyes of hatred.",
-            "An angry roar of fury is released at {target}, with throbbing veins of anger from {user}.",
-            "{user} yells a string of menacing curses at {target}, with a filled voice of ire.",
-            "{user} unleashes a powerful scream of rage at {target}, with a twisted face of anger.",
-            "A stream of violent curses is screamed out at {target}, with a voice filled with venom from {user}.",
-            "{user} shouts a tirade of expletives at {target}, with throbbing veins of anger.",
-            "A violent bellow of anger is released at {target}, with bulging eyes of rage from {user}.",
-            "{user} growls a stream of expletives at {target}, with a reddening face of anger."
-        ],
-        self: [
-            "Yelling at yourself won't make your life better, you poor thing.",
-            "What an idiot, trying to yell himself out of depression, ha!",
-            "Yup, you are an idiot, trying to yell at yourself to make everything better, who do you think you are?",
-            "You think yelling at yourself would make you feel better? Wow, nobody even knows how much of a loser you are.",
-            "Yelling at yourself, what sort of madness is that? Find somebody else to yell at, you pathetic crybaby.",
-            "Hey, don't you think it's kinda pointless to yell at yourself? Nobody even knows how desperate you are.",
-            "How sad is your life, that even yelling at yourself doesn't make you feel better.",
-            "Yelling at yourself, oof, nobody even knows how lonely you are, nobody would even care.",
-            "Yelling at yourself won't get you anywhere, why don't you try finding somebody to talk to?",
-            "Maybe if you stop yelling at yourself and find somebody to confide in it'll make you feel better.",
-            "Makes sense that you're trying to yell at yourself, nobody even knows how sad and pathetic you are.",
-            "Yelling at yourself, you think that'll get you anywhere? Wow, nobody knows how stupid and desperate you are.",
-            "What a joke, you can't even yell at someone else and now you're trying to yell at yourself, unbelievable.",
-            "Yelling at yourself, oh wow, that's sad. Nobody even knows how vulnerable and insecure you are.",
-            "Y'know what? Maybe if you stop yelling at yourself and find somebody else to talk to it'll help a bit.",
-            "Yelling at yourself, so you're mad at someone else but you can't say it out loud, got it.",
-            "Yelling at yourself? Wow, why don't you stop and try seeking help, that nobody knows how uncool and depressed you are.",
-            "How sad are you, that you are so desperate to make yourself feel better, you resort to yelling at yourself?",
-            "Yelling at yourself won't help, maybe if you reach out to someone it'll make you feel better, brave man.",
-            "Why don't you stop yelling at yourself and start looking for help, yeah? That'll make you feel better.",
-            "Yelling at yourself isn't gonna help, why don't you try doing something with your time, something productive? That'll make you feel better."
-
-        ],
-        bot: [
-            "Yell at me one more time and I'm gonna fucking kill you!",
-            "I'm gonna break your face if you keep yelling at me.",
-            "What the fuck is wrong with you? Stop yelling at me!",
-            "If you think you can just come up and yell at me, you're wrong.",
-            "You're lucky I don't punch you right now. Stop yelling at me.",
-            "I'm gonna make you sorry if you keep yelling at me.",
-            "You want to yell? Come get some.",
-            "I'm gonna kick your ass if you try to yell at me again.",
-            "I swear to God, if you don't stop yelling at me I'm gonna fight you.",
-            "I'm gonna knock you the fuck out if you ever try to yell at me again.",
-            "You think you can just yell at me? Good luck with that.",
-            "You're a fucking idiot, you know that? You can't yell at me, you moron.",
-            "I'm not scared of your yelling, so give it up or else.",
-            "I'm gonna break your arm if you try to yell at me again.",
-        ]
-    }
-}
-
-function getRandomItem(arr: string[]): string {
-    return arr[Math.floor(Math.random() * arr.length)];
-}
-
-export function randomFeeling(
-    category: keyof typeof Feelings,
-    group: keyof FeelingsData
-): string {
-    const feelingCategory = Feelings[category] as FeelingsData;
-    const selectedGroup = feelingCategory[group];
-
-    if (!selectedGroup || selectedGroup.length === 0) {
-        throw new Error(`No data found for group: ${group} in category: ${category}`);
-    }
-
-    return getRandomItem(selectedGroup);
-}
